Move stories to dedicated directory
