--- src/components/Icons.stories.tsx
@@ -1,261 +1,0 @@
-import ArrowTopRightIcon from './icons/ArrowTopRightIcon'
-import BotIcon from './icons/BotIcon'
-import BrowserIcon from './icons/BrowserIcon'
-import BundleIcon from './icons/BundleIcon'
-import CaretDownIcon from './icons/CaretDownIcon'
-import CertificateIcon from './icons/CertificateIcon'
-import CheckIcon from './icons/CheckIcon'
-import ChecklistIcon from './icons/ChecklistIcon'
-import ChronjobIcon from './icons/ChronjobIcon'
-import CloseIcon from './icons/CloseIcon'
-import ClusterIcon from './icons/ClusterIcon'
-import CollapseIcon from './icons/CollapseIcon'
-import ComponentsIcon from './icons/ComponentsIcon'
-import ComputerNodeIcon from './icons/ComputerNodeIcon'
-import CopyIcon from './icons/CopyIcon'
-import CraneIcon from './icons/CraneIcon'
-import CreditCardIcon from './icons/CreditCardIcon'
-import DashboardIcon from './icons/DashboardIcon'
-import DeploymentIcon from './icons/DeploymentIcon'
-import DiscordIcon from './icons/DiscordIcon'
-import DockerTagIcon from './icons/DockerTagIcon'
-import DomainsIcon from './icons/DomainsIcon'
-import DownloadIcon from './icons/DownloadIcon'
-import DropdownArrowIcon from './icons/DropdownArrowIcon'
-import EditIcon from './icons/EditIcon'
-import EmojiHoverIcon from './icons/EmojiHoverIcon'
-import EmojiIcon from './icons/EmojiIcon'
-import ErrorIcon from './icons/ErrorIcon'
-import EyeClosedIcon from './icons/EyeClosedIcon'
-import EyeIcon from './icons/EyeIcon'
-import FileIcon from './icons/FileIcon'
-import FiltersIcon from './icons/FiltersIcon'
-import FingerPrintIcon from './icons/FingerPrintIcon'
-import FolderIcon from './icons/FolderIcon'
-import GearTrainIcon from './icons/GearTrainIcon'
-import GitHubIcon from './icons/GitHubIcon'
-import GitHubLogoIcon from './icons/GitHubLogoIcon'
-import GlobeIcon from './icons/GlobeIcon'
-import GraphIcon from './icons/GraphIcon'
-import HamburgerMenuCollapseIcon from './icons/HamburgerMenuCollapseIcon'
-import HamburgerMenuCollapsedIcon from './icons/HamburgerMenuCollapsedIcon'
-import HamburgerMenuIcon from './icons/HamburgerMenuIcon'
-import HistoryIcon from './icons/HistoryIcon'
-import IdIcon from './icons/IdIcon'
-import InstallIcon from './icons/InstallIcon'
-import InstalledIcon from './icons/InstalledIcon'
-import InvoicesIcon from './icons/InvoicesIcon'
-import KeyIcon from './icons/KeyIcon'
-import KeyPairIcon from './icons/KeyPairIcon'
-import LifePreserverIcon from './icons/LifePreserverIcon'
-import LightningIcon from './icons/LightningIcon'
-import LinkoutIcon from './icons/LinkoutIcon'
-import LinksIcon from './icons/LinksIcon'
-import ListIcon from './icons/ListIcon'
-import LogoutIcon from './icons/LogoutIcon'
-import LogsIcon from './icons/LogsIcon'
-import MagnifyingGlassIcon from './icons/MagnifyingGlassIcon'
-import MailIcon from './icons/MailIcon'
-import MarketIcon from './icons/MarketIcon'
-import MarketPlusIcon from './icons/MarketPlusIcon'
-import MessagesIcon from './icons/MessagesIcon'
-import NetworkInIcon from './icons/NetworkInIcon'
-import NetworkInterfaceIcon from './icons/NetworkInterfaceIcon'
-import NewMailIcon from './icons/NewMailIcon'
-import OAuthIcon from './icons/OAuthIcon'
-import OwnerIcon from './icons/OwnerIcon'
-import PackageIcon from './icons/PackageIcon'
-import PadlockIcon from './icons/PadlockIcon'
-import PadlockLockedIcon from './icons/PadlockLockedIcon'
-import PaperclipIcon from './icons/PaperclipIcon'
-import PencilIcon from './icons/PencilIcon'
-import PeopleIcon from './icons/PeopleIcon'
-import PeoplePlusIcon from './icons/PeoplePlusIcon'
-import PersonIcon from './icons/PersonIcon'
-import PersonPlusIcon from './icons/PersonPlusIcon'
-import PlanIcon from './icons/PlanIcon'
-import PushPinIcon from './icons/PushPinIcon'
-import ReloadIcon from './icons/ReloadIcon'
-import RocketIcon from './icons/RocketIcon'
-import RunBookIcon from './icons/RunBookIcon'
-import ScrollIcon from './icons/ScrollIcon'
-import ScrollPlusIcon from './icons/ScrollPlusIcon'
-import SendMessageIcon from './icons/SendMessageIcon'
-import ServersIcon from './icons/ServersIcon'
-import ShipIcon from './icons/ShipIcon'
-import SirenIcon from './icons/SirenIcon'
-import SortAscIcon from './icons/SortAscIcon'
-import SortDescIcon from './icons/SortDescIcon'
-import StatusIpIcon from './icons/StatusIpIcon'
-import StatusOkIcon from './icons/StatusOkIcon'
-import SuitcaseIcon from './icons/SuitcaseIcon'
-import TableIcon from './icons/TableIcon'
-import TagIcon from './icons/TagIcon'
-import TerminalIcon from './icons/TerminalIcon'
-import TrashCanIcon from './icons/TrashCanIcon'
-import TwitterIcon from './icons/TwitterIcon'
-import UpdatesIcon from './icons/UpdatesIcon'
-import VolumesIcon from './icons/VolumesIcon'
-import WebhooksIcon from './icons/WebhooksIcon'
-
-const icons = {
-  ArrowTopRightIcon,
-  BotIcon,
-  BrowserIcon,
-  BundleIcon,
-  CaretDownIcon,
-  CertificateIcon,
-  CheckIcon,
-  ChecklistIcon,
-  ChronjobIcon,
-  CloseIcon,
-  ClusterIcon,
-  CollapseIcon,
-  ComponentsIcon,
-  ComputerNodeIcon,
-  CopyIcon,
-  CraneIcon,
-  CreditCardIcon,
-  DashboardIcon,
-  DeploymentIcon,
-  DiscordIcon,
-  DockerTagIcon,
-  DomainsIcon,
-  DownloadIcon,
-  DropdownArrowIcon,
-  EditIcon,
-  EmojiHoverIcon,
-  EmojiIcon,
-  ErrorIcon,
-  EyeClosedIcon,
-  EyeIcon,
-  FileIcon,
-  FiltersIcon,
-  FingerPrintIcon,
-  FolderIcon,
-  GearTrainIcon,
-  GitHubIcon,
-  GitHubLogoIcon,
-  GlobeIcon,
-  GraphIcon,
-  HamburgerMenuCollapseIcon,
-  HamburgerMenuCollapsedIcon,
-  HamburgerMenuIcon,
-  HistoryIcon,
-  IdIcon,
-  InstallIcon,
-  InstalledIcon,
-  InvoicesIcon,
-  KeyIcon,
-  KeyPairIcon,
-  LifePreserverIcon,
-  LightningIcon,
-  LinkoutIcon,
-  LinksIcon,
-  ListIcon,
-  LogoutIcon,
-  LogsIcon,
-  MagnifyingGlassIcon,
-  MailIcon,
-  MarketIcon,
-  MarketPlusIcon,
-  MessagesIcon,
-  NetworkInIcon,
-  NetworkInterfaceIcon,
-  NewMailIcon,
-  OAuthIcon,
-  OwnerIcon,
-  PackageIcon,
-  PadlockIcon,
-  PadlockLockedIcon,
-  PaperclipIcon,
-  PencilIcon,
-  PeopleIcon,
-  PeoplePlusIcon,
-  PersonIcon,
-  PersonPlusIcon,
-  PlanIcon,
-  PushPinIcon,
-  ReloadIcon,
-  RocketIcon,
-  RunBookIcon,
-  ScrollIcon,
-  ScrollPlusIcon,
-  SendMessageIcon,
-  ServersIcon,
-  ShipIcon,
-  SirenIcon,
-  SortAscIcon,
-  SortDescIcon,
-  StatusIpIcon,
-  StatusOkIcon,
-  SuitcaseIcon,
-  TableIcon,
-  TagIcon,
-  TerminalIcon,
-  TrashCanIcon,
-  TwitterIcon,
-  UpdatesIcon,
-  VolumesIcon,
-  WebhooksIcon,
-}
-
-export default {
-  title: 'Icons',
-  component: MarketPlusIcon,
-}
-
-function Template(args: any) {
-  return (
-    <div style={{
-      display: 'flex',
-      flexWrap: 'wrap',
-      maxWidth: '100%',
-    }}
-    >
-      {Object.entries(icons).map(([name, Icon]) => (
-        <div
-          key={name}
-          style={{
-            padding: '0.25rem',
-            marginBottom: '1.5rem',
-            marginRight: '1rem',
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
-            width: 32 + 16,
-          }}
-        >
-          <span style={{ marginBottom: '0.5rem', fontSize: '0.75rem' }}>
-            {name.replace('Icon', '')}
-          </span>
-          <Icon {...args} />
-        </div>
-      ))}
-    </div>
-  )
-}
-
-export const Default = Template.bind({})
-Default.args = {
-}
-
-export const Xlarge = Template.bind({})
-Xlarge.args = {
-  size: 32,
-}
-
-export const Large = Template.bind({})
-Large.args = {
-  size: 24,
-}
-
-export const Small = Template.bind({})
-Small.args = {
-  size: 12,
-}
-
-export const Color = Template.bind({})
-Color.args = {
-  color: 'primary',
-}
