+++ src/stories/Alert.stories.tsx
@@ -1,4 +1,4 @@
-import Alert from './Alert'
+import Alert from '../components/Alert'
 
 export default {
   title: 'Alert',
