+++ src/stories/Icons.stories.tsx
@@ -1,0 +1,261 @@
+import ArrowTopRightIcon from '../components/icons/ArrowTopRightIcon'
+import BotIcon from '../components/icons/BotIcon'
+import BrowserIcon from '../components/icons/BrowserIcon'
+import BundleIcon from '../components/icons/BundleIcon'
+import CaretDownIcon from '../components/icons/CaretDownIcon'
+import CertificateIcon from '../components/icons/CertificateIcon'
+import CheckIcon from '../components/icons/CheckIcon'
+import ChecklistIcon from '../components/icons/ChecklistIcon'
+import ChronjobIcon from '../components/icons/ChronjobIcon'
+import CloseIcon from '../components/icons/CloseIcon'
+import ClusterIcon from '../components/icons/ClusterIcon'
+import CollapseIcon from '../components/icons/CollapseIcon'
+import ComponentsIcon from '../components/icons/ComponentsIcon'
+import ComputerNodeIcon from '../components/icons/ComputerNodeIcon'
+import CopyIcon from '../components/icons/CopyIcon'
+import CraneIcon from '../components/icons/CraneIcon'
+import CreditCardIcon from '../components/icons/CreditCardIcon'
+import DashboardIcon from '../components/icons/DashboardIcon'
+import DeploymentIcon from '../components/icons/DeploymentIcon'
+import DiscordIcon from '../components/icons/DiscordIcon'
+import DockerTagIcon from '../components/icons/DockerTagIcon'
+import DomainsIcon from '../components/icons/DomainsIcon'
+import DownloadIcon from '../components/icons/DownloadIcon'
+import DropdownArrowIcon from '../components/icons/DropdownArrowIcon'
+import EditIcon from '../components/icons/EditIcon'
+import EmojiHoverIcon from '../components/icons/EmojiHoverIcon'
+import EmojiIcon from '../components/icons/EmojiIcon'
+import ErrorIcon from '../components/icons/ErrorIcon'
+import EyeClosedIcon from '../components/icons/EyeClosedIcon'
+import EyeIcon from '../components/icons/EyeIcon'
+import FileIcon from '../components/icons/FileIcon'
+import FiltersIcon from '../components/icons/FiltersIcon'
+import FingerPrintIcon from '../components/icons/FingerPrintIcon'
+import FolderIcon from '../components/icons/FolderIcon'
+import GearTrainIcon from '../components/icons/GearTrainIcon'
+import GitHubIcon from '../components/icons/GitHubIcon'
+import GitHubLogoIcon from '../components/icons/GitHubLogoIcon'
+import GlobeIcon from '../components/icons/GlobeIcon'
+import GraphIcon from '../components/icons/GraphIcon'
+import HamburgerMenuCollapseIcon from '../components/icons/HamburgerMenuCollapseIcon'
+import HamburgerMenuCollapsedIcon from '../components/icons/HamburgerMenuCollapsedIcon'
+import HamburgerMenuIcon from '../components/icons/HamburgerMenuIcon'
+import HistoryIcon from '../components/icons/HistoryIcon'
+import IdIcon from '../components/icons/IdIcon'
+import InstallIcon from '../components/icons/InstallIcon'
+import InstalledIcon from '../components/icons/InstalledIcon'
+import InvoicesIcon from '../components/icons/InvoicesIcon'
+import KeyIcon from '../components/icons/KeyIcon'
+import KeyPairIcon from '../components/icons/KeyPairIcon'
+import LifePreserverIcon from '../components/icons/LifePreserverIcon'
+import LightningIcon from '../components/icons/LightningIcon'
+import LinkoutIcon from '../components/icons/LinkoutIcon'
+import LinksIcon from '../components/icons/LinksIcon'
+import ListIcon from '../components/icons/ListIcon'
+import LogoutIcon from '../components/icons/LogoutIcon'
+import LogsIcon from '../components/icons/LogsIcon'
+import MagnifyingGlassIcon from '../components/icons/MagnifyingGlassIcon'
+import MailIcon from '../components/icons/MailIcon'
+import MarketIcon from '../components/icons/MarketIcon'
+import MarketPlusIcon from '../components/icons/MarketPlusIcon'
+import MessagesIcon from '../components/icons/MessagesIcon'
+import NetworkInIcon from '../components/icons/NetworkInIcon'
+import NetworkInterfaceIcon from '../components/icons/NetworkInterfaceIcon'
+import NewMailIcon from '../components/icons/NewMailIcon'
+import OAuthIcon from '../components/icons/OAuthIcon'
+import OwnerIcon from '../components/icons/OwnerIcon'
+import PackageIcon from '../components/icons/PackageIcon'
+import PadlockIcon from '../components/icons/PadlockIcon'
+import PadlockLockedIcon from '../components/icons/PadlockLockedIcon'
+import PaperclipIcon from '../components/icons/PaperclipIcon'
+import PencilIcon from '../components/icons/PencilIcon'
+import PeopleIcon from '../components/icons/PeopleIcon'
+import PeoplePlusIcon from '../components/icons/PeoplePlusIcon'
+import PersonIcon from '../components/icons/PersonIcon'
+import PersonPlusIcon from '../components/icons/PersonPlusIcon'
+import PlanIcon from '../components/icons/PlanIcon'
+import PushPinIcon from '../components/icons/PushPinIcon'
+import ReloadIcon from '../components/icons/ReloadIcon'
+import RocketIcon from '../components/icons/RocketIcon'
+import RunBookIcon from '../components/icons/RunBookIcon'
+import ScrollIcon from '../components/icons/ScrollIcon'
+import ScrollPlusIcon from '../components/icons/ScrollPlusIcon'
+import SendMessageIcon from '../components/icons/SendMessageIcon'
+import ServersIcon from '../components/icons/ServersIcon'
+import ShipIcon from '../components/icons/ShipIcon'
+import SirenIcon from '../components/icons/SirenIcon'
+import SortAscIcon from '../components/icons/SortAscIcon'
+import SortDescIcon from '../components/icons/SortDescIcon'
+import StatusIpIcon from '../components/icons/StatusIpIcon'
+import StatusOkIcon from '../components/icons/StatusOkIcon'
+import SuitcaseIcon from '../components/icons/SuitcaseIcon'
+import TableIcon from '../components/icons/TableIcon'
+import TagIcon from '../components/icons/TagIcon'
+import TerminalIcon from '../components/icons/TerminalIcon'
+import TrashCanIcon from '../components/icons/TrashCanIcon'
+import TwitterIcon from '../components/icons/TwitterIcon'
+import UpdatesIcon from '../components/icons/UpdatesIcon'
+import VolumesIcon from '../components/icons/VolumesIcon'
+import WebhooksIcon from '../components/icons/WebhooksIcon'
+
+const icons = {
+  ArrowTopRightIcon,
+  BotIcon,
+  BrowserIcon,
+  BundleIcon,
+  CaretDownIcon,
+  CertificateIcon,
+  CheckIcon,
+  ChecklistIcon,
+  ChronjobIcon,
+  CloseIcon,
+  ClusterIcon,
+  CollapseIcon,
+  ComponentsIcon,
+  ComputerNodeIcon,
+  CopyIcon,
+  CraneIcon,
+  CreditCardIcon,
+  DashboardIcon,
+  DeploymentIcon,
+  DiscordIcon,
+  DockerTagIcon,
+  DomainsIcon,
+  DownloadIcon,
+  DropdownArrowIcon,
+  EditIcon,
+  EmojiHoverIcon,
+  EmojiIcon,
+  ErrorIcon,
+  EyeClosedIcon,
+  EyeIcon,
+  FileIcon,
+  FiltersIcon,
+  FingerPrintIcon,
+  FolderIcon,
+  GearTrainIcon,
+  GitHubIcon,
+  GitHubLogoIcon,
+  GlobeIcon,
+  GraphIcon,
+  HamburgerMenuCollapseIcon,
+  HamburgerMenuCollapsedIcon,
+  HamburgerMenuIcon,
+  HistoryIcon,
+  IdIcon,
+  InstallIcon,
+  InstalledIcon,
+  InvoicesIcon,
+  KeyIcon,
+  KeyPairIcon,
+  LifePreserverIcon,
+  LightningIcon,
+  LinkoutIcon,
+  LinksIcon,
+  ListIcon,
+  LogoutIcon,
+  LogsIcon,
+  MagnifyingGlassIcon,
+  MailIcon,
+  MarketIcon,
+  MarketPlusIcon,
+  MessagesIcon,
+  NetworkInIcon,
+  NetworkInterfaceIcon,
+  NewMailIcon,
+  OAuthIcon,
+  OwnerIcon,
+  PackageIcon,
+  PadlockIcon,
+  PadlockLockedIcon,
+  PaperclipIcon,
+  PencilIcon,
+  PeopleIcon,
+  PeoplePlusIcon,
+  PersonIcon,
+  PersonPlusIcon,
+  PlanIcon,
+  PushPinIcon,
+  ReloadIcon,
+  RocketIcon,
+  RunBookIcon,
+  ScrollIcon,
+  ScrollPlusIcon,
+  SendMessageIcon,
+  ServersIcon,
+  ShipIcon,
+  SirenIcon,
+  SortAscIcon,
+  SortDescIcon,
+  StatusIpIcon,
+  StatusOkIcon,
+  SuitcaseIcon,
+  TableIcon,
+  TagIcon,
+  TerminalIcon,
+  TrashCanIcon,
+  TwitterIcon,
+  UpdatesIcon,
+  VolumesIcon,
+  WebhooksIcon,
+}
+
+export default {
+  title: 'Icons',
+  component: MarketPlusIcon,
+}
+
+function Template(args: any) {
+  return (
+    <div style={{
+      display: 'flex',
+      flexWrap: 'wrap',
+      maxWidth: '100%',
+    }}
+    >
+      {Object.entries(icons).map(([name, Icon]) => (
+        <div
+          key={name}
+          style={{
+            padding: '0.25rem',
+            marginBottom: '1.5rem',
+            marginRight: '1rem',
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'center',
+            width: 32 + 16,
+          }}
+        >
+          <span style={{ marginBottom: '0.5rem', fontSize: '0.75rem' }}>
+            {name.replace('Icon', '')}
+          </span>
+          <Icon {...args} />
+        </div>
+      ))}
+    </div>
+  )
+}
+
+export const Default = Template.bind({})
+Default.args = {
+}
+
+export const Xlarge = Template.bind({})
+Xlarge.args = {
+  size: 32,
+}
+
+export const Large = Template.bind({})
+Large.args = {
+  size: 24,
+}
+
+export const Small = Template.bind({})
+Small.args = {
+  size: 12,
+}
+
+export const Color = Template.bind({})
+Color.args = {
+  color: 'primary',
+}
