+++ src/stories/Tag.stories.tsx
@@ -1,4 +1,4 @@
-import Tag from './Tag'
+import Tag from '../components/Tag'
 
 export default {
   title: 'Tag',
