+++ src/stories/UserCard.stories.tsx
@@ -1,4 +1,4 @@
-import UserCard from './UserCard'
+import UserCard from '../components/UserCard'
 
 export default {
   title: 'UserCard',
